+++ packages/provider/lib/src/async_provider.dart
@@ -17,8 +17,15 @@ typedef ErrorBuilder<T> = T Function(BuildContext context, Object error);
 
 /// Listens to a [Stream<T>] and exposes [T] to its descendants.
 ///
-/// It is considered an error to pass a stream that can emit errors without providing
-/// a [catchError] method.
+/// Its main use-case it to provide to a large number of a widget the content
+/// of an existing stream, without caring about reacting to events.
+///
+/// A typical example would be to expose the battery level, or a Firebase query.
+/// Trying to use [Stream] to replace [ChangeNotifier] is outside of the scope
+/// of this class.
+///
+/// It is considered an error to pass a stream that can emit errors without
+/// providing a [catchError] method.
 ///
 /// {@template provider.streamprovider.initialdata}
 /// [initialData] determines the value exposed until the [Stream] emits a value.
@@ -54,8 +61,8 @@ class StreamProvider<T> extends ValueDelegateWidget<Stream<T>>
 
   /// Creates a [StreamController] from [builder] and subscribes to its stream.
   ///
-  /// [StreamProvider] will automatically call [StreamController.close]
-  /// when the widget is removed from the tree.
+  /// [StreamProvider] will automatically call [StreamController.close] when the
+  /// widget is removed from the tree.
   ///
   /// The parameter [builder] must not be `null`.
   StreamProvider.controller({
@@ -110,11 +117,11 @@ class StreamProvider<T> extends ValueDelegateWidget<Stream<T>>
 
   /// An optional function used whenever the [Stream] emits an error.
   ///
-  /// [catchError] will be called with the emitted error and
-  /// is expected to return a fallback value without throwing.
+  /// [catchError] will be called with the emitted error and is expected to
+  /// return a fallback value without throwing.
   ///
-  /// The returned value will then be exposed to the descendants of [StreamProvider]
-  /// like any valid value.
+  /// The returned value will then be exposed to the descendants of
+  /// [StreamProvider] like any valid value.
   final ErrorBuilder<T> catchError;
 
   /// {@macro provider.updateshouldnotify}
@@ -200,8 +207,8 @@ class _StreamControllerBuilderDelegate<T>
 
 /// Listens to a [Future<T>] and exposes [T] to its descendants.
 ///
-/// It is considered an error to pass a future that can emit errors without providing
-/// a [catchError] method.
+/// It is considered an error to pass a future that can emit errors without
+/// providing a [catchError] method.
 ///
 /// {@macro provider.updateshouldnotify}
 ///
@@ -267,11 +274,11 @@ class FutureProvider<T> extends ValueDelegateWidget<Future<T>>
 
   /// Optional function used if the [Future] emits an error.
   ///
-  /// [catchError] will be called with the emitted error and
-  /// is expected to return a fallback value without throwing.
+  /// [catchError] will be called with the emitted error and is expected to
+  /// return a fallback value without throwing.
   ///
-  /// The returned value will then be exposed to the descendants of [FutureProvider]
-  /// like any valid value.
+  /// The returned value will then be exposed to the descendants of
+  /// [FutureProvider] like any valid value.
   final ErrorBuilder<T> catchError;
 
   /// {@macro provider.updateshouldnotify}