+++ packages/provider/lib/src/provider.dart
@@ -109,8 +109,8 @@ class _InheritedProviderElement<T> extends InheritedElement {
 }
 
 /// A provider that merges multiple providers into a single linear widget tree.
-/// It is used to improve readability and reduce boilderplate code of having to
-/// nest mutliple layers of providers.
+/// It is used to improve readability and reduce boilerplate code of having to
+/// nest multiple layers of providers.
 ///
 /// As such, we're going from:
 ///
@@ -203,8 +203,10 @@ class MultiProvider extends StatelessWidget
 /// [Provider] is removed from the tree.
 ///
 /// {@template provider.updateshouldnotify}
-/// [updateShouldNotify] can optionally be passed to avoid unnecessaryly rebuilding dependants when nothing changed.
-/// Defaults to `(previous, next) => previous != next`. See [InheritedWidget.updateShouldNotify] for more informations.
+/// [updateShouldNotify] can optionally be passed to avoid unnecessarily
+/// rebuilding dependents when nothing changed. Defaults to
+/// `(previous, next) => previous != next`. See
+/// [InheritedWidget.updateShouldNotify] for more information.
 /// {@endtemplate}
 ///
 /// ```dart
@@ -223,9 +225,39 @@ class MultiProvider extends StatelessWidget
 ///   }
 /// }
 /// ```
+///
+/// ## Testing
+///
+/// When testing widgets that consumes providers, it is necessary to
+/// add the proper providers in the widget tree above the tested widget.
+///
+/// A typical test may like like this:
+///
+/// ```dart
+/// final foo = MockFoo();
+///
+/// await tester.pumpWidget(
+///   Provider<Foo>.value(
+///     value: foo,
+///     child: TestedWidget(),
+///   ),
+/// );
+/// ```
+///
+/// Note this example purposefully specified the object type, instead of having
+/// it infered.
+/// Since we used a mocked class (typically using `mockito`), then we have to
+/// downcast the mock to the type of the mocked class.
+/// Otherwise, the type inference will resolve to `Provider<MockFoo>` instead of
+/// `Provider<Foo>`, which will cause `Provider.of<Foo>` to fail.
 class Provider<T> extends ValueDelegateWidget<T>
     implements SingleChildCloneableWidget {
-  /// Allows to specify parameters to [Provider].
+  /// Creates a value, store it, and expose it to its descendants.
+  ///
+  /// The value can be optionally disposed using [dispose] callback. This
+  /// callbackwhich will be called when [Provider] is unmounted from the
+  /// widget tree, or if [Provider] is rebuilt to use [Provider.value] instead.
+  ///
   Provider({
     Key key,
     @required ValueBuilder<T> builder,
@@ -269,7 +301,8 @@ class Provider<T> extends ValueDelegateWidget<T>
     this.child,
   }) : super(key: key, delegate: delegate);
 
-  /// Obtains the nearest [Provider<T>] up its widget tree and returns its value.
+  /// Obtains the nearest [Provider<T>] up its widget tree and returns its
+  /// value.
   ///
   /// If [listen] is `true` (default), later value changes will trigger a new
   /// [State.build] to widgets, and [State.didChangeDependencies] for
@@ -291,14 +324,15 @@ class Provider<T> extends ValueDelegateWidget<T>
     return element._value;
   }
 
-  /// A sanity check to prevent misuse of [Provider] when a variant should be used.
+  /// A sanity check to prevent misuse of [Provider] when a variant should be
+  /// used.
   ///
-  /// By default, [debugCheckInvalidValueType] will throw if `value` is a [Listenable]
-  /// or a [Stream].
-  /// In release mode, [debugCheckInvalidValueType] does nothing.
+  /// By default, [debugCheckInvalidValueType] will throw if `value` is a
+  /// [Listenable] or a [Stream].  In release mode, [debugCheckInvalidValueType]
+  /// does nothing.
   ///
-  /// This check can be disabled altogether by setting [debugCheckInvalidValueType]
-  /// to `null` like so:
+  /// This check can be disabled altogether by setting
+  /// [debugCheckInvalidValueType] to `null` like so:
   ///
   /// ```dart
   /// void main() {
@@ -314,7 +348,7 @@ Tried to use Provider with a subtype of Listenable/Stream ($T).
 
 This is likely a mistake, as Provider will not automatically update dependents
 when $T is updated. Instead, consider changing Provider for more specific
-implementation that handles the update mecanism, such as:
+implementation that handles the update mechanism, such as:
 
 - ListenableProvider
 - ChangeNotifierProvider